Update and rename functions.v1.2.R to functions.v1.3.R
Flexibility was added to ms() function, by removing the argument | tail +4 | grep -v // that was giving troubles to linux users. This command will clean up the ms output for gene tree simulation (by removing a header and all the "//" separating loci). 
Now this is done with R, and should solve the problem.